chore: refactor ClassInterest
