+++ src/features/ClassInterest/components/ClassInterestHeader.tsx
@@ -1,0 +1,11 @@
+export default function ClassInterestHeader() {
+	return (
+		<div>
+			<h2 className="text-2xl font-bold mb-2">Class Interests</h2>
+			<p>
+				Select and reorder your preferred WoW classes. Use the arrow buttons to
+				reorder selected items.
+			</p>
+		</div>
+	);
+}
